Reuse the same `CUDA_HOME` logic from cuDF
Checks for `cuda-gdb`, which is still visible when the Conda `nvcc` shim
package is installed. Also errors out if `CUDA_HOME` cannot be found.
Includes a clear error message as to what needs to be done to fix it.

Same code is in cuDF. So this is really just copying that logic.