Construct logger directly to avoid registration.
Using the spdlog factories registers the logger
with the given name. It is an error to try and
register more than one logger with the same name
which caused a problem of trying to create
more than one `logging_resource_adaptor` in the
same application (since the name is fixed). Therefore,
avoid using the factories all together and just manually
create the logger object.