cmake: add CNMEM_INCLUDE_DIR as a private include for rmm
This is all a bit iffy, since we're compiling an outside source (cnmem.cpp)
into librmm, and that expects to find cnmem.h in a certain path, which
is provided here. For users, however, that cnmem.h is copied into
rmm/detail/cnmem.h, which is in the public include path for librmm.