cmake: mv thirdparty/CMakeLists.txt -> cmake/RMM_thirdparty.cmake
The main thing here is that the file is now included, rather than
pulled in via add_subdirectory(). The intended consequence of that is
that it's not executed in a separate scope, so that `find_package()`,
or rather its wrapped form `CPMFindPackage()` will work properly.

The one change after the move is removing the `PARENT_SCOPE` from setting
`THRUST_INCLUDE_DIR`, since it's now running in the parent scope.