cmake: prefer locally installed thirdparty packages (#572)
* cmake: mv thirdparty/CMakeLists.txt -> cmake/RMM_thirdparty.cmake

The main thing here is that the file is now included, rather than
pulled in via add_subdirectory(). The intended consequence of that is
that it's not executed in a separate scope, so that `find_package()`,
or rather its wrapped form `CPMFindPackage()` will work properly.

The one change after the move is removing the `PARENT_SCOPE` from setting
`THRUST_INCLUDE_DIR`, since it's now running in the parent scope.

* cmake: support locally installed spdlog

This will prefer using a locally installed version of spdlog -- only if that
cannot be found will it download its own copy.

* cmake: support locally installed googletest

This works relatively easily, since new googletest comes with its own
cmake config. In the case where CPM pulls it in via add_subdirectory,
we have to create appropriate aliases to the new namespaced names.

* cmake: support locally installed googlebenchmark