Avoid potential race conditions in device_scalar/device_uvector setters (#725)
There is a subtle problem with `rmm::device_uvector::set_element_async` and `rmm::device_scalar::set_value`. It's common to pass a literal (e.g. 0) to these functions. But these functions accept the parameter by reference, so if you pass a literal, it's possible for the temporary created to store the literal to be destroyed before the `cudaMemcpyAsync` is performed, resulting in a use after free.  This PR:

1. Disallows passing literals to these functions by deleting the rvalue reference (`&&`) overload. This is a breaking API change.
2. In `device_scalar`, adds an optimization for `bool` types to always use `cudaMemsetAsync`. 
3. In `device_scalar`, add a new method `set_value_zero` for the common case of initialization to zero. Also uses `cudaMemsetAsync`.
4. Improves documentation.

These changes will require PRs to fix up some uses in cudf, cuspatial, and cugraph.

Authors:
  - Mark Harris (@harrism)

Approvers:
  - Rong Ou (@rongou)
  - Conor Hoekstra (@codereport)

URL: https://github.com/rapidsai/rmm/pull/725