Execution policy class (#816)
Changes currently planned for the RAFT code include updating the RAFT handle for it to store a singleton of a thrust execution policy. An execution policy class would prove useful to write a clearer code. This PR is a proof of concept for it.

cc @divyegala

Authors:
  - Victor Lafargue (https://github.com/viclafargue)

Approvers:
  - Divye Gala (https://github.com/divyegala)
  - Mark Harris (https://github.com/harrism)

URL: https://github.com/rapidsai/rmm/pull/816