RMM shouldn't install thrust
When we merged https://github.com/rapidsai/rmm/pull/854 we enabled
the installation of Thrust as part of RMM install. This causes
downstream issues as projects like cuDF want to use a patched
version to improve compile times.

Since RMM is header only we don't need to ensure our consumers
use the same Thrust version, so we can skip installing it.