Add the ability to register and unregister reinitialization hooks (#1072)
This PR introduces the ability to register one or more "hooks" that will be invoked before each call to `rmm.reinitialize()`.

The motivation is to allow downstream libraries to register hooks that are responsible for deleting any internal objects that that keep references to an RMM memory resource. In particular, this is useful for https://github.com/rapidsai/cudf/pull/11246.

Authors:
  - Ashwin Srinath (https://github.com/shwina)

Approvers:
  - Vyas Ramasubramani (https://github.com/vyasr)
  - https://github.com/jakirkham
  - Peter Andreas Entschev (https://github.com/pentschev)
  - Bradley Dice (https://github.com/bdice)
  - Lawrence Mitchell (https://github.com/wence-)

URL: https://github.com/rapidsai/rmm/pull/1072