Address alignment differences with `cuda::mr::resource_ref`
Our implementation of `cuda::mr::{async_}resource_ref` currently aligns to `alignof(max_align_t)`

This is a minor difference to the current device_memory_resource. We intent to expand the interface of `cuda::mr::{async_}resource_ref` to take desired alignment into account.

However, this is out of scope of this PR and will be done in the cccl repository