Accept stream argument in DeviceMemoryResource allocate/deallocate
In #980, the DeviceMemoryResource class in Python gained allocation
and deallocation routines. This was to facility writing Python
allocate/deallocate callbacks for the CallbackMemoryResource.

These routines should, to match the C++ API, accept a stream parameter
such that one can use them for stream-ordered allocation. Although we
recommend that users allocate on the Python side using the
DeviceBuffer interface, exposing these routines implicitly makes them
public.

To fix this, add an optional stream argument defaulting to the default
stream.

- Closes #1493