Fix search path for torch allocator in editable installs and ensure CUDA support is available (#1498)
The current logic for finding the allocator .so file presumes that torch.py is in the same directory tree as the compiled allocator. While this is true when the package is installed, it is not true for editable installs of the package where py files are left in the source tree while compiled modules are placed into a build tree (both are found by indirection).

The current allocator logic also assumes that if pytorch is available it is a version compiled with CUDA support. This PR also adds a check to verify that we don't try to set the allocator on a CPU-only build of torch.

Authors:
  - Vyas Ramasubramani (https://github.com/vyasr)
  - Bradley Dice (https://github.com/bdice)

Approvers:
  - Lawrence Mitchell (https://github.com/wence-)
  - Mark Harris (https://github.com/harrism)

URL: https://github.com/rapidsai/rmm/pull/1498