DeviceBuffer: accept memory resource when taking ownership
In c_from_unique_ptr we should not just rely on
get_current_device_resource, but rather allow the user to pass in the
memory resource they _know_ was used to allocate the buffer we are
taking ownership of.

So that we are backwards-compatible we default, as before, to the
current device resource.