Cargo-cult a Doxygen setup that correctly expands macros
It appears to be impossible to get doxygen to EXPAND_AS_DEFINED a
macro that is in a file that is excluded via EXCLUDE_PATTERNS (even if
it is explicitly listed in INPUT). If the file is only excluded via
EXCLUDE, then it is visible to the preprocessor, but that setting
overrides any INPUT (so explicitly listed files that are EXCLUDEd are
no longer documented).

Consequently, we must unfortunately just repeat the preprocessor
definitions for RMM_NAMESPACE, RMM_EXPORT, and RMM_HIDDEN.