Sccache stats telemetry (#1830)
This pr uses new functionality added to shared-actions and shared-workflows to capture sccache hit rate information. To add this to other repos, we'll need to make the slight alteration here:

`sccache --show-adv-stats | tee ../../telemetry-artifacts/sccache-stats.txt`

That is, output the sccache stats to a file with a particular name in the telemetry-artifacts folder.

Authors:
  - Mike Sarahan (https://github.com/msarahan)

Approvers:
  - James Lamb (https://github.com/jameslamb)

URL: https://github.com/rapidsai/rmm/pull/1830