Feat/more none checking (#2120)
This PR adds a utility to validate that input Stream arguments are not None. In the future it can be generalized to support other input types that support the CUDA stream protocol.

Note that while Cython supports a `not None` syntax for function parameters, [that only works for `def` functions](https://github.com/cython/cython/issues/3329), so it is not suitable here since we need to support `cdef` and `cpdef` functions as well. Also, since we will eventually want to support the stream protocol we'd be removing that annotation eventually.

Authors:
  - Vyas Ramasubramani (https://github.com/vyasr)
  - Bradley Dice (https://github.com/bdice)

Approvers:
  - Lawrence Mitchell (https://github.com/wence-)
  - Bradley Dice (https://github.com/bdice)

URL: https://github.com/rapidsai/rmm/pull/2120