Use PyData docs theme, reorganize C++ docs (#2137)
This is split off of #2087.

I am overhauling the RMM documentation. This is the first set of changes, which includes a new theme and a reorganization of the C++ docs. All docs now use Markdown / Myst.

The next phases will include docstring tweaks to fix various formatting/cross-linking issues (see #2138 and #2139 for current progress on this), an expansion of the Python API docs, and adding user guides for various features.

Authors:
  - Bradley Dice (https://github.com/bdice)

Approvers:
  - Matthew Murray (https://github.com/Matt711)
  - Rong Ou (https://github.com/rongou)
  - Jake Awe (https://github.com/AyodeAwe)

URL: https://github.com/rapidsai/rmm/pull/2137