Remove upper bound on numpy dependency
This removes the upper bound on our numpy dependency.

The primary motivation is to avoid allowing prereleases in the project's
metadata. Including the `<3.0.0a` allows pip to consider pre-release
packages in its resolution, so pip installing RMM will currently bring
in a pre-release version of numpy by default:

```
$ python -m pip install --dry-run --extra-index-url https://pypi.anaconda.org/rapidsai-wheels-nightly/simple "rmm-cu12>=26.2.0a0" "librmm-cu12>=26.2.0a0"
...
Would install cuda-bindings-12.9.4 cuda-pathfinder-1.3.3 cuda-python-12.9.4 librmm-cu12-26.2.0a22 numpy-2.4.0rc1 rapids-logger-0.2.3 rmm-cu12-26.2.0a22
```

despite the user not explicitly requesting a prerelase NumPy, or
globally allowing prereleases.

To solve this, we could either remove the `a0` in the upper bound,
or remove the upper bound entirely. I've opted to remove it entirely

1. NumPy doesn't follow semver. API breaking changes are made in "minor"
   releases (after deprecation). "Major" release are typically used
   for ABI breaking changes. I don't *think* RMM uses the numpy C API,
   I might be wrong.
2. I wasn't too clear on the original motivation for the upper bound.
   This was originally added in https://github.com/rapidsai/rmm/pull/1501 and
   then bumped to `<3.0.0a0` in https://github.com/rapidsai/rmm/pull/1650.

I'm happy to adjust the to just `<3.0.0` if the upper bound is
preferred.